# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CycleCare API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: Get pad requests and allocation summary for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Requests summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestsSummary"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new pad request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PadRequest"
        "400":
          description: Invalid input or allocation exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Active request already exists this month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests/{id}/collect:
    patch:
      operationId: markRequestCollected
      tags: [requests]
      summary: Mark a pad request as collected
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request marked collected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PadRequest"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PadRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        quantity:
          type: integer
        status:
          type: string
          enum: [pending, approved, ready, collected, rejected]
        pickupCode:
          type: string
        pickupLocation:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - quantity
        - status
        - pickupCode
        - pickupLocation
        - createdAt
        - updatedAt
    RequestsSummary:
      type: object
      properties:
        allocation:
          type: object
          properties:
            total:
              type: integer
            used:
              type: integer
            remaining:
              type: integer
          required:
            - total
            - used
            - remaining
        currentRequest:
          oneOf:
            - $ref: "#/components/schemas/PadRequest"
            - type: "null"
        history:
          type: array
          items:
            $ref: "#/components/schemas/PadRequest"
      required:
        - allocation
        - currentRequest
        - history
    CreateRequestInput:
      type: object
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 25
        notes:
          type: ["string", "null"]
      required:
        - quantity
...